achieve the left of index page

+++ src/main/java/com/community/cyd/dto/GithubUser.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class GithubUser {
     private String name;// 用户名
-    private Long id;    // 用户id
+    private Integer id;    // 用户id
     private String bio; // 描述
     private String avatarUrl; //头像地址
 }
